backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
httpx==0.28.1
python-multipart==0.0.20
pypdf==5.9.0
python-docx==1.2.0
python-dotenv==1.1.1
reportlab==4.4.3
jinja2==3.1.6
playwright==1.58.0
